pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "browsecomp-v3"
version = "0.1.0"
description = "Browsecomp风格复杂学术问题生成器"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Hu Family", email = "your-email@example.com"}
]
keywords = ["knowledge-graph", "question-generation", "academic-qa"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "networkx>=3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "rich>=13.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "mypy>=1.5",
    "ruff>=0.1.0",
]
llm = [
    "openai>=1.0",
    "langchain>=0.1",
]
viz = [
    "matplotlib>=3.7",
    "graphviz>=0.20",
]

[project.scripts]
browsecomp = "browsecomp_v3.main:main"

[project.urls]
Homepage = "https://github.com/your-username/browsecomp-V3"
Repository = "https://github.com/your-username/browsecomp-V3"
Documentation = "https://github.com/your-username/browsecomp-V3/docs"

[tool.setuptools.packages.find]
where = ["."]
include = ["browsecomp_v3*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
